+++ lib/Elastica/Result.php
@@ -186,6 +186,24 @@ class Result
     {
         return $this->getParam('_explanation');
     }
+    
+    /**
+     * Returns Document
+     * 
+     * @return \Elastica\Document
+     */
+    public function getDocument()
+    {
+        $doc = new \Elastica\Document();
+        $doc->setData($this->getSource());
+        $hit = $this->getHit();
+        if($this->hasParam('_source')) unset($hit['_source']);
+        if($this->hasParam('_explanation')) unset($hit['_explanation']);
+        if($this->hasParam('highlight')) unset($hit['highlight']);
+        if($this->hasParam('_score')) unset($hit['_score']);
+        $doc->setParams($hit);
+        return $doc;
+    }
 
     /**
      * Magic function to directly access keys inside the result.